requirements.txt
fastapi
jinja2
uvicorn
openai
ipykernel
python-multipart
ffmpeg-python